requirements.txt
google-generativeai
flask
flask-socketio